refactor: delete CoreModal, this will also fixed #362
