+++ docs/content/2.get-started/1.guide/2.setup.md
@@ -60,7 +60,7 @@ export default defineNuxtPlugin((nuxtApp) => {
 
 ## Import required CSS
 
-vue-final-modal 4 has tiny size of required CSS (gzipped 0.49kb). All classes have a `.vfm-` prefix, so you don't have to worry about any CSS pollution. _[See style of `<VueFinalModal>`.](https://github.com/vue-final/vue-final-modal/blob/v4/packages/vue-final-modal/src/components/CoreModal/CoreModal.vue#L184-L217)_
+vue-final-modal 4 has tiny size of required CSS (gzipped 0.49kb). All classes have a `.vfm-` prefix, so you don't have to worry about any CSS pollution.
 
 ### Vue 3
