+++ packages/vue-final-modal/cypress/components/useZIndex.spec.ts
@@ -25,24 +25,22 @@ describe('Test useZIndex()', () => {
         stubs: { transition: false },
       },
     }).as('app')
-    cy.get('@app').then(() => {
-      firstModal.open()
-    })
+
+    cy.get('@app').then(() => firstModal.open())
     cy.get('.first-modal').should(($el) => {
       expect($el).to.have.css('zIndex', '1000')
     })
-    cy.get('@app').then(() => {
-      secondModal.open()
-    })
+
+    cy.get('@app').then(() => secondModal.open())
     cy.get('.second-modal').should(($el) => {
       expect($el).to.have.css('zIndex', '1002')
     })
-    cy.get('@app').then(() => {
-      thirdModal.open()
-    })
+
+    cy.get('@app').then(() => thirdModal.open())
     cy.get('.third-modal').should(($el) => {
       expect($el).to.have.css('zIndex', '1004')
     })
+
     cy.get('@app').then(() => {
       thirdModal.patchOptions({
         attrs: {
@@ -53,9 +51,8 @@ describe('Test useZIndex()', () => {
     cy.get('.third-modal').should(($el) => {
       expect($el).to.have.css('zIndex', '1238')
     })
-    cy.get('@app').then(() => {
-      firstModal.close()
-    })
+
+    cy.get('@app').then(() => firstModal.close())
     cy.get('.second-modal').should(($el) => {
       expect($el).to.have.css('zIndex', '1000')
     })
