+++ packages/vue-final-modal/src/components/VueFinalModal/VueFinalModalProps.ts
@@ -1,8 +1,12 @@
-import type { PropType, RendererElement } from 'vue'
-import { coreModalProps } from '../CoreModal/CoreModalProps'
+import type { Options } from 'focus-trap'
+import type { PropType, RendererElement, TransitionProps } from 'vue'
+import type { ModalId, StyleValue } from '~/Modal'
+
+// Hack from: https://github.com/microsoft/TypeScript/issues/29729#issuecomment-1331857805
+type AnyString = string & {}
+type VfmTransition = 'vfm-fade' | 'vfm-slide-down' | 'vfm-slide-up' | 'vfm-slide-right' | 'vfm-slide-left' | AnyString
 
 export const vueFinalModalProps = {
-  ...coreModalProps,
   /**
    * @description Set `null | false` to disable teleport.
    * @default `'body'`
@@ -14,5 +18,227 @@ export const vueFinalModalProps = {
   teleportTo: {
     type: [String, null, Boolean, Object] as PropType<string | RendererElement | null | false>,
     default: 'body',
+  },
+  /**
+   * @description An uniq name for the open/close a modal via vfm.open/vfm.close APIs.
+   * @default `undefined`
+   * @example Symbol: `Symbol('MyModal')`
+   * @example String: `'AUniqString'`
+   * @example Number: `300`
+   */
+  modalId: {
+    type: [String, Number, Symbol] as PropType<ModalId>,
+    default: undefined,
+  },
+  /**
+   * @description Display the modal or not.
+   * @default `undefined`
+   * @example
+   * ```js
+   * const showModal = ref(false)
+   * v-model="showModal"
+   * ```
+   */
+  modelValue: {
+    type: Boolean as PropType<boolean>,
+    default: undefined,
+  },
+  /**
+   * @description Render the modal via `if` or `show`.
+   * @default `'if'`
+   * @example
+   * ```js
+   * displayDirective: 'if'
+   * ```
+   * @example
+   * ```js
+   * displayDirective: 'show'
+   * ```
+   */
+  displayDirective: {
+    type: String as PropType<'if' | 'show' | 'visible'>,
+    default: 'if',
+    validator: (prop: any) => ['if', 'show', 'visible'].includes(prop),
+  },
+  /**
+   * @description Hide the overlay or not.
+   * @default `undefined`
+   * @example
+   * ```js
+   * hideOverlay="true"
+   * ```
+   */
+  hideOverlay: {
+    type: Boolean as PropType<boolean>,
+    default: undefined,
+  },
+  /**
+   * @description Customize the overlay transition.
+   * @default `undefined`
+   */
+  overlayTransition: {
+    type: [String, Object] as PropType<VfmTransition | TransitionProps>,
+    default: undefined,
+  },
+  /**
+   * @description Customize the content transition.
+   * @default `undefined`
+   */
+  contentTransition: {
+    type: [String, Object] as PropType<VfmTransition | TransitionProps>,
+    default: undefined,
+  },
+  /**
+   * @description Bind class to vfm__overlay.
+   * @default `undefined`
+   */
+  overlayClass: {
+    type: undefined as unknown as PropType<any>,
+    default: undefined,
+  },
+  /**
+   * @description Bind class to vfm__content.
+   * @default `undefined`
+   */
+  contentClass: {
+    type: undefined as unknown as PropType<any>,
+    default: undefined,
+  },
+  /**
+   * @description Bind style to vfm__overlay.
+   * @default `undefined`
+   */
+  overlayStyle: {
+    type: [String, Object, Array] as PropType<StyleValue>,
+    default: undefined,
+  },
+  /**
+   * @description Bind style to vfm__content.
+   * @default `undefined`
+   */
+  contentStyle: {
+    type: [String, Object, Array] as PropType<StyleValue>,
+    default: undefined,
+  },
+  /**
+   * @description Is it allow to close the modal by clicking the overlay.
+   * @default `true`
+   */
+  clickToClose: {
+    type: Boolean as PropType<boolean>,
+    default: true,
+  },
+  /**
+   * @description Is it allow to close the modal by keypress `esc`.
+   * @default `true`
+   */
+  escToClose: {
+    type: Boolean as PropType<boolean>,
+    default: true,
+  },
+  /**
+   * @description Is it allow to click outside of the vfm__content when the modal is opened
+   * @default `'non-interactive'`
+   */
+  background: {
+    type: String as PropType<'interactive' | 'non-interactive'>,
+    default: 'non-interactive',
+    validator: (prop: any) => ['interactive', 'non-interactive'].includes(prop),
+  },
+  /**
+   * @description
+   * * Use `{ disabled: true }` to disable the focusTrap.
+   * * Checkout the createOptions type here https://github.com/focus-trap/focus-trap for more.
+   * @default `{ allowOutsideClick: true }`
+   */
+  focusTrap: {
+    type: [Boolean, Object] as PropType<false | Options>,
+    default: () => ({
+      allowOutsideClick: true,
+    }),
+  },
+  /**
+   * @description Lock body scroll or not when the modal is opened.
+   * @default `true`
+   */
+  lockScroll: {
+    type: Boolean as PropType<boolean>,
+    default: true,
+  },
+  /**
+   * @description Creates a padding-right when scroll is locked to prevent the page from jumping
+   * @default `true`
+   */
+  reserveScrollBarGap: {
+    type: Boolean as PropType<boolean>,
+    default: true,
+  },
+  /**
+   * @description Define how to increase the zIndex when there are nested modals
+   * @default `({ index }) => 1000 + 2 * index`
+   */
+  zIndexFn: {
+    type: Function as PropType<(context: { index: number }) => number | undefined>,
+    default: ({ index }: { index: number }) => 1000 + 2 * index,
+  },
+  /**
+   * @description The direction of swiping to close the modal
+   * @default `none`
+   * @example
+   * Set swipeToClose="none" to disable swiping to close
+   * ```js
+   * swipeToClose="none"
+   * ```
+   */
+  swipeToClose: {
+    type: String as PropType<'none' | 'up' | 'right' | 'down' | 'left'>,
+    default: 'none',
+    validator: (prop: any) => ['none', 'up', 'right', 'down', 'left'].includes(prop),
+  },
+  /**
+   * @description Threshold for swipe to close
+   * @default `0`
+   */
+  threshold: {
+    type: Number as PropType<number>,
+    default: 0,
+  },
+  /**
+   * @description If set `:showSwipeBanner="true"`, only allow clicking `swipe-banner` slot to swipe to close
+   * @default `undefined`
+   * @example
+   * ```js
+   * swipeToClose="right"
+   * :showSwipeBanner="true"
+   * ```
+   * ```html
+   * <VueFinalModal
+   *   ...
+   *   swipeToClose="right"
+   *   :showSwipeBanner="true"
+   * >
+   *   <template #swipe-banner>
+   *     <div style="position: absolute; height: 100%; top: 0; left: 0; width: 10px;" />
+   *   </template>
+   *   ...modal content
+   * </VueFinalModal>
+   * ```
+   */
+  showSwipeBanner: {
+    type: Boolean as PropType<boolean>,
+    default: undefined,
+  },
+  /**
+   * @description When set `:preventNavigationGestures="true"`, there will be two invisible bars for prevent navigation gestures including swiping back/forward on mobile webkit. For example: Safari mobile.
+   * @default `undefined`
+   * @example
+   * Set preventNavigationGestures="true" to prevent Safari navigation gestures including swiping back/forward.
+   * ```js
+   * :preventNavigationGestures="true"
+   * ```
+   */
+  preventNavigationGestures: {
+    type: Boolean as PropType<boolean>,
+    default: undefined,
   },
 } as const
