+++ packages/vue-final-modal/src/components/VueFinalModal/useModelValue.ts
@@ -1,10 +1,10 @@
 import { nextTick, ref, watch } from 'vue'
 import type { Ref } from 'vue'
-import type CoreModal from './CoreModal.vue'
+import type VueFinalModal from './VueFinalModal.vue'
 
 export function useModelValue(
-  props: InstanceType<typeof CoreModal>['$props'],
-  emit: InstanceType<typeof CoreModal>['$emit'],
+  props: InstanceType<typeof VueFinalModal>['$props'],
+  emit: InstanceType<typeof VueFinalModal>['$emit'],
   options: {
     open: () => boolean
     close: () => boolean
