+++ packages/vue-final-modal/src/components/VueFinalModal/useTransition.ts
@@ -1,6 +1,6 @@
 import type { ComputedRef, Ref, TransitionProps } from 'vue'
 import { computed, nextTick, ref, watch } from 'vue'
-import type CoreModal from './CoreModal.vue'
+import type VueFinalModal from './VueFinalModal.vue'
 
 export enum TransitionState {
   Enter,
@@ -31,7 +31,7 @@ function useTransitionState(_visible = false): [Ref<boolean>, Ref<undefined | Tr
 }
 
 export function useTransition(
-  props: InstanceType<typeof CoreModal>['$props'],
+  props: InstanceType<typeof VueFinalModal>['$props'],
   options: {
     modelValueLocal: Ref<boolean>
     onEntering?: () => void
