+++ packages/vue-final-modal/src/components/VueFinalModal/vVisible.ts
