+++ packages/vue-final-modal/src/index.ts
@@ -1,5 +1,4 @@
 import ModalsContainer from './components/ModalsContainer.vue'
-import CoreModal from './components/CoreModal/CoreModal.vue'
 import VueFinalModal from './components/VueFinalModal/VueFinalModal.vue'
 
 import type { Vfm } from './Modal'
@@ -13,11 +12,9 @@ export * from './plugin'
 /** Components */
 export {
   ModalsContainer,
-  CoreModal,
   VueFinalModal,
 }
 
-export { coreModalProps } from './components/CoreModal/CoreModalProps'
 export { vueFinalModalProps } from './components/VueFinalModal/VueFinalModalProps'
 
 export type { VueFinalModalEmits } from './components/VueFinalModal/VueFinalModal.vue'
